Algunos ejercicios corregidos Acceso Datos

+++ AccesoADatos/src/tema1/ejercicios/_01_01/ejercicio2.java
@@ -1,4 +1,4 @@
-package tema1.ejercicios;
+package tema1.ejercicios._01_01;
 
 import java.io.*;
 import java.util.Scanner;
@@ -11,12 +11,11 @@ public class ejercicio2 {
 		System.out.println("Introduce el directorio");
 		File directorio = new File(scanner.nextLine());
 
-		if (directorio.exists()) {
+		if (directorio.exists() && directorio.isFile()) {
 			System.out.println("Directorio borrado");
 			directorio.delete(); // Solo lo borra si esta vacio
 
 		} else {
-			directorio.mkdir();
 			System.out.println("El directorio no existe");
 		}
